+++ homeassistant/components/iqvia/manifest.json
@@ -1,0 +1,13 @@
+{
+  "domain": "iqvia",
+  "name": "IQVIA",
+  "documentation": "https://www.home-assistant.io/components/iqvia",
+  "requirements": [
+    "numpy==1.16.2",
+    "pyiqvia==0.2.0"
+  ],
+  "dependencies": [],
+  "codeowners": [
+    "@bachya"
+  ]
+}
