--- homeassistant/components/pollen/__init__.py
@@ -1,1 +1,0 @@
-"""The pollen component."""
